+++ app/src/main/assets/changelog.md
@@ -1,5 +1,6 @@
 ### 1.0.3-beta
 * **New:** Calculate file MD5 & SHA1
+* **Fix:** Copy rclone and config file in the background
 
 ***